[CSL-2205] Get rid of Block.Network.Announce module
